+++ src/app/(route)/main/_components/live/NewPost.tsx
