+++ src/app/(route)/main/_components/live/NewPostItem.tsx
